pyproject.toml
[project]
name = "zu.pdresume"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ZackaryW", email = "36378555+ZackaryW@users.noreply.github.com" }
]
dependencies = [
    "zuu>=8.5.0",
    "pyyaml>=6.0.2",
    "toml>=0.10.2",
    "click>=8.1.8",
]
readme = "README.md"
requires-python = ">= 3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/zu"]

[project.scripts]
zu-pdresume = "zu.pdresume.__main__:cli"